Move FaDInLoglikelihood and FaDInNoPrecomputations to experiments
